requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
argostranslate==1.9.6
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ctranslate2==4.3.1
dataclasses-json==0.6.7
datasets==2.20.0
decorator==5.1.1
dill==0.3.8
et-xmlfile==1.1.0
evaluate==0.4.2
executing==2.0.1
faiss-cpu==1.8.0.post1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.1
grpcio-status==1.62.2
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.24.0
idna==3.7
intel-openmp==2021.4.0
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.9
langchain-community==0.2.7
langchain-core==0.2.21
langchain-google-genai==1.0.7
langchain-text-splitters==0.2.2
langsmith==0.1.92
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
openpyxl==3.1.5
orjson==3.10.6
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pdf2image==1.17.0
pillow==10.4.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
pure-eval==0.2.2
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.1.2
PyPDF2==3.0.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
sacremoses==0.0.53
selenium==4.22.0
sentencepiece==0.2.0
setuptools==71.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
stanza==1.1.1
streamlit==1.36.0
sympy==1.13.1
tbb==2021.13.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
trio==0.26.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
watchdog==4.0.1
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
Levenshtein
